refactor: 장바구니 아이템 로직 커스텀훅으로 분리
